Reworked the 'join' and 'part' commands to be compatible with relay groups. Specification of channel AND network to join or part remains the same. However, if you pass only a single argument 'join' or 'part' it is assumed to be a relay group and, if valid, all of the channels in the group will be join'ed or part'ed.
Also added an 'add' command which allow a channel to be added to an existing
relay group, or if the specified relay group doesn't exist, a new one is
created.  Syntax for this is:

/add #channel net group rcv xmit password

These arguments correspond to the elements in the %Relays structure in the
config file.  The last three arguments are optional in the sense that if they
are unspecified, default values of 1, 1, and '' will be assumed.  However,
to specify a password, for example, rcv and xmit must be specified.

Once a channel has been added to a group, it is not automatically 'join'ed.
It is merely eligible to be 'join'ed the next time a 'join' request is issued
to that channel or group.